src/backend/requirements.txt
# FastAPI and ASGI server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database and ORM
sqlmodel>=0.0.14
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Authentication
python-jose[cryptography]>=3.3.0

# Form handling
python-multipart>=0.0.6

# Environment variables
python-dotenv>=1.0.0

# PostgreSQL driver (psycopg2-binary works best with Vercel)
psycopg2-binary>=2.9.9

# Email validation
email-validator>=2.1.0

# Testing (optional - can be removed for production)
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.26.0

# Phase III: AI & Agents
openai>=1.12.0
tenacity>=8.2.3  # Retry logic for OpenAI API
# Note: openai-agents-sdk and mcp-sdk are not yet available as packages
# Will use openai SDK directly for agent implementation
